+++ pyagentspec/src/pyagentspec/adapters/langgraph/_types.py
@@ -15,11 +15,12 @@ if TYPE_CHECKING:
     # Otherwise, importing the module when they are not installed would lead to an import error.
 
     import langchain.agents as langchain_agents
+    import langchain.agents.middleware.types as langchain_middleware_types
     import langchain_ollama
     import langchain_openai
     import langgraph.graph as langgraph_graph
     import langgraph_swarm
-    from langchain.agents.middleware.types import AgentState
+    from langchain.agents.middleware.types import AgentMiddleware, AgentState
     from langchain_core.callbacks import BaseCallbackHandler
     from langchain_core.language_models import BaseChatModel
     from langchain_core.messages import BaseMessage, SystemMessage, ToolMessage
@@ -39,6 +40,7 @@ else:
     langchain_openai = LazyLoader("langchain_openai")
     langgraph_graph = LazyLoader("langgraph.graph")
     langchain_agents = LazyLoader("langchain.agents")
+    langchain_middleware_types = LazyLoader("langchain.agents.middleware.types")
     BaseTool = LazyType("langchain_core.tools", "BaseTool")
     StructuredTool = LazyType("langchain_core.tools", "StructuredTool")
     Checkpointer = LazyType("langgraph.types", "Checkpointer")
@@ -59,6 +61,7 @@ else:
     GenerationChunk = LazyType("langchain_core.outputs", "GenerationChunk")
     LLMResult = LazyType("langchain_core.outputs", "LLMResult")
     AgentState = LazyType("langchain.agents.middleware.types", "AgentState")
+    AgentMiddleware = LazyType("langchain.agents.middleware.types", "AgentMiddleware")
 
 
 LangGraphTool: TypeAlias = Union[BaseTool, Callable[..., Any]]
@@ -108,6 +111,7 @@ ControlFlow: TypeAlias = Dict[SourceNodeId, Dict[BranchName, TargetNodeId]]
 __all__ = [
     "langgraph_graph",
     "langchain_agents",
+    "langchain_middleware_types",
     "langchain_ollama",
     "langchain_openai",
     "LangGraphTool",
@@ -136,6 +140,7 @@ __all__ = [
     "ToolMessage",
     "BaseChatModel",
     "AgentState",
+    "AgentMiddleware",
     "Checkpointer",
     "interrupt",
     "RunnableConfig",